nestest OK

+++ emulator-go/processor/nrom.go
@@ -3,16 +3,16 @@ package processor
 import "log"
 
 type Mapper interface {
-	Read(address int) byte
-	Write(address int, data byte)
+	Read(address uint16) byte
+	Write(address uint16, data byte)
 }
 
 type NROM struct {
 	*Cartridge
-	banks int
+	banks uint16
 }
 
-func (m NROM) Read(address int) byte {
+func (m NROM) Read(address uint16) byte {
 	switch {
 	case address < 0x2000:
 		return m.CHR[address]
@@ -31,12 +31,12 @@ func (m NROM) Read(address int) byte {
 	return 0
 }
 
-func (m NROM) Write(address int, data byte) {
+func (m NROM) Write(address uint16, data byte) {
 	switch {
 	case address < 0x2000:
 		m.CHR[address] = data
 	case address >= 0x8000:
-		log.Printf("NROM write %02X (%d)", data, int(data)%m.banks)
+		log.Printf("NROM write %02X (%d)", data, uint16(data)%m.banks)
 	case address >= 0x6000:
 		if m.Battery {
 			index := address - 0x6000
@@ -50,7 +50,7 @@ func (m NROM) Write(address int, data byte) {
 func CreateMapper(cartridge *Cartridge) Mapper {
 	switch cartridge.Mapper {
 	case 0x00:
-		return NROM{cartridge, len(cartridge.ROM) / 0x4000}
+		return NROM{cartridge, uint16(len(cartridge.ROM) / 0x4000)}
 	default:
 		return nil
 	}
